Updated Python Code with reactances and taxonomies
This new version incorporates new taxonomies, such as "reduction substation" and "boundary substation", but they have the same parameters as "Substation". The difference is whether they play the role of source or consumer node (or both), which is not relevant for assigning different fragility functions. 

New line properties: "Resistance" and "Reactance". They are important for computing the weight for the load and capacity (Origin-Destination betweenness centrality) estimation.

To Dos: In next version the role of the nodes will be indicated in a new node property (source, consumer, transmission, source and consumer), and the taxonomies will be closer taxonomies to the FEMA framework (substations by voltages and power plants by sizes)