Changed the factor of max_dist from 2.1 to 55.61
the former parameter was too restrictive for the locations of the nodal components in the system_reliability process. In particular, many nodes read zero intensity, even if located inside the extent of the shakemap. Therefore, understanding that the distance is given in km and the long and lat spacing are given in degrees, the new factor makes max_dist to be half the arc length with angle equal to the hypothenuse of lon_spacing and lat_spacing, assuming the Earth radius equal to 6372.82 km. The factor also includes the conversion from degrees to radians for the arc length formula.